sentinella producer tier + degradata (built, measured, dormant)
Codex 61: negative constant returned through a cast from an
unsigned function - the tcp/tls disease at birth; cast-only (bare
form is parity 54's). First harvest: sputnik unknown escapes
emitted truncated 0xFF (producer + three dead >= 0 consumers,
VANA tail 46->43); sententia_iterator_index sentinel now s32.
Codex 62 (unsigned <= 0): census 228 = the emptiness idiom owns
the shape; needle needs def-use. Dormant behind _degradata_vigil
with the re-arm condition named for CFG stage 2.

+++ lib/sententia_fissio.c
@@ -1290,16 +1290,16 @@ sententia_iterator_prior(
 }
 
 
-i32
+s32
 sententia_iterator_index(
     SententiaIterator* iter)
 {
     si (iter == NIHIL)
     {
-        redde (i32)-1;
+        redde -1;
     }
 
-    redde iter->index_currens;
+    redde (s32)iter->index_currens;
 }
